// packages/shared/generated/types.ts
export interface paths {
// ...
    "/info": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Get solution information */
        get: {
            parameters: {
                query?: {
                    /** @description Filter by filename */
                    file?: components["parameters"]["FileFilter"];
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Solution metadata and statistics */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["InfoResponse"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/get": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Get object by UUID */
        get: {
            parameters: {
                query: {
                    /** @description Object UUID */
                    uuid: string;
                    /**
                     * @description Output format
                     * @default json
                     */
                    format?: components["parameters"]["Format"];
                    /** @description Include SQL query in response */
                    debug?: components["parameters"]["Debug"];
                    /** @description Include metadata in response */
                    meta?: components["parameters"]["Meta"];
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Object details */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ObjectResponse"];
                    };
                };
                404: components["responses"]["NotFound"];
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/list": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** List objects by type */
        get: {
            parameters: {
                query: {
                    /** @description Object type to list */
                    type: components["schemas"]["ObjectType"];
                    /** @description Filter by filename */
                    file?: components["parameters"]["FileFilter"];
                    /** @description Maximum number of results (0 = all) */
                    limit?: components["parameters"]["Limit"];
                    /**
                     * @description Output format
                     * @default json
                     */
                    format?: components["parameters"]["Format"];
                    /** @description Include SQL query in response */
                    debug?: components["parameters"]["Debug"];
                    /** @description Include metadata in response */
                    meta?: components["parameters"]["Meta"];
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description List of objects */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ObjectListResponse"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/search": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Search objects by name */
        get: {
            parameters: {
                query: {
                    /** @description Name pattern (SQL LIKE, e.g. %Import%) */
                    name: string;
                    type?: components["schemas"]["ObjectType"];
                    /** @description Filter by filename */
                    file?: components["parameters"]["FileFilter"];
                    /** @description Maximum number of results (0 = all) */
                    limit?: components["parameters"]["Limit"];
                    /** @description Number of results to skip (for pagination) */
                    offset?: components["parameters"]["Offset"];
                    /**
                     * @description Output format
                     * @default json
                     */
                    format?: components["parameters"]["Format"];
                    /** @description Include SQL query in response */
                    debug?: components["parameters"]["Debug"];
                    /** @description Include metadata in response */
                    meta?: components["parameters"]["Meta"];
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Search results */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ObjectListResponse"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/search/count": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Count search results by name pattern */
        get: {
            parameters: {
                query: {
                    /** @description Name pattern (SQL LIKE, e.g. %Import%) */
                    name: string;
                    type?: components["schemas"]["ObjectType"];
                    /** @description Filter by filename */
                    file?: components["parameters"]["FileFilter"];
                    /**
                     * @description Output format
                     * @default json
                     */
                    format?: components["parameters"]["Format"];
                    /** @description Include SQL query in response */
                    debug?: components["parameters"]["Debug"];
                    /** @description Include metadata in response */
                    meta?: components["parameters"]["Meta"];
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Search result count */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["CountResponse"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/count": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Count objects */
        get: {
            parameters: {
                query?: {
                    type?: components["schemas"]["ObjectType"];
                    /** @description Filter by filename */
                    file?: components["parameters"]["FileFilter"];
                    /** @description Group results by field */
                    group_by?: "type" | "file" | "type,file";
                    /**
                     * @description Output format
                     * @default json
                     */
                    format?: components["parameters"]["Format"];
                    /** @description Include SQL query in response */
                    debug?: components["parameters"]["Debug"];
                    /** @description Include metadata in response */
                    meta?: components["parameters"]["Meta"];
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Object counts */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["CountResponse"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/references": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Get object references */
        get: {
            parameters: {
                query: {
                    /** @description Object UUID */
                    uuid: string;
                    /** @default all */
                    direction?: components["schemas"]["ReferenceDirection"];
                    /** @default operational */
                    link_type?: components["schemas"]["LinkType"];
                    /** @description Maximum number of results (0 = all) */
                    limit?: components["parameters"]["Limit"];
                    /**
                     * @description Output format
                     * @default json
                     */
                    format?: components["parameters"]["Format"];
                    /** @description Include SQL query in response */
                    debug?: components["parameters"]["Debug"];
                    /** @description Include metadata in response */
                    meta?: components["parameters"]["Meta"];
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Object references */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ReferencesResponse"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        ResponseMeta: {
            /** @description Query execution time in milliseconds */
            execution_time_ms?: number;
            /** @description Number of results returned */
            result_count?: number;
            /** @description SQL query (only if debug=true) */
            query?: string;
        };
// ...
    };
// ...
}